requirements.txt
amqp==1.4.9
anyjson==0.3.3
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.23
celery==3.1.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.16
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-guardian==1.4.9
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
Flask==1.0.2
idna==2.7
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10
kombu==3.0.37
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==1.3.0
numpy==1.7.1
postgres==2.2.2
psycopg2==2.7.6.1
psycopg2-binary==2.7.5
PyJWT==1.6.4
pyparsing==1.5.7
python-dateutil==2.7.3
pytz==2018.7
requests==2.19.1
six==1.11.0
tornado==4.0
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
Werkzeug==0.14.1
